feat: Using double in AABBTree x3

+++ Allay-Server/src/test/java/cn/allay/api/datastruct/aabbtree/TestEntity.java
@@ -4,25 +4,25 @@ import org.joml.primitives.AABBd;
 
 public class TestEntity implements HasAABB, HasLongId {
 
-   private final float x;
+   private final double x;
 
-   private final float y;
+   private final double y;
 
-   private final float z;
+   private final double z;
 
-   private final float width;
+   private final double width;
 
-   private final float height;
+   private final double height;
 
-    private final float length;
+    private final double length;
 
    private final int id;
 
-   public TestEntity(int id, float x, float y, float width, float height) {
+   public TestEntity(int id, double x, double y, double width, double height) {
         this(id, x, y, 0, width, height, 0);
    }
 
-   public TestEntity(int id, float x, float y, float z, float width, float height, float length) {
+   public TestEntity(int id, double x, double y, double z, double width, double height, double length) {
       this.id = id;
       this.x = x;
       this.y = y;
